refactor: split index into modular components
