+++ .gitignore
@@ -1,0 +1,5 @@
+node_modules
+.env
+dist
+uploads
+
