+++ api.ts
@@ -1,0 +1,1 @@
+export const API_BASE = import.meta.env.VITE_API_URL || '';
