+++ basic.test.js
@@ -1,0 +1,6 @@
+import { test } from 'node:test';
+import assert from 'node:assert';
+
+test('dummy test passes', () => {
+  assert.strictEqual(1 + 1, 2);
+});
